:recycle: refactor(DrugsTable): Split and make it a function component.
Progress on #213, #214, #303, #304
