+++ imports/api/publication/drugs/find.ts
@@ -1,9 +1,8 @@
 import {Drugs} from '../../collection/drugs';
+import pageQuery from '../../pageQuery';
 import define from '../define';
 
 export default define({
 	name: 'drugs',
-	cursor() {
-		return Drugs.find({owner: this.userId});
-	},
+	cursor: pageQuery(Drugs),
 });
