+++ imports/ui/drugs/DrugsTable.tsx
@@ -1,0 +1,1 @@
+export {default} from '../table/EnhancedTable';
